Moved GetData() and GetDataFromFile to MerinoCore so we can use it in other windows.
