+++ Assets/Merino/Editor/Scripts/MerinoEditorWindow.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
 using System.IO;
@@ -7,14 +6,11 @@ using System.Linq;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
-using Yarn;
 using File = System.IO.File;
-using Merino.EditorCoroutines;
 // wow that's a lot of usings
 
 namespace Merino
 {
-
 	class MerinoEditorWindow : EditorWindow
 	{
 		// sidebar tree view management stuff
@@ -164,7 +160,7 @@ namespace Merino
 				multiColumnHeader.ResizeToFit ();
 
 			// generate sidebar data structures
-			var treeModel = new TreeModel<MerinoTreeElement>(GetData(null, true));
+			var treeModel = new TreeModel<MerinoTreeElement>(MerinoCore.GetData());
 			m_TreeView = new MerinoTreeView(MerinoData.ViewState, multiColumnHeader, treeModel);
 			m_TreeView.treeChanged += OnTreeChanged;
 			
@@ -317,7 +313,7 @@ namespace Merino
 					// remove project folder path from selected path (+1 = trailing slash)
 					addFilePath = "Assets/" + addFilePath.Substring(Application.dataPath.Length + 1);
 					LoadYarnFileAtFullPath(addFilePath, true); // add all found files to currentFiles
-					m_TreeView.treeModel.SetData(GetData());
+					m_TreeView.treeModel.SetData(MerinoCore.GetData());
 					m_TreeView.Reload();
 				}
 			}
@@ -337,7 +333,7 @@ namespace Merino
 					// remove project folder path from selected path (+1 = trailing slash)
 					addPath = addPath == Application.dataPath ? "Assets" : "Assets/" + addPath.Substring( Application.dataPath.Length + 1);
 					LoadYarnFilesAtPath(addPath); // add all found files to currentFiles
-					m_TreeView.treeModel.SetData(GetData());
+					m_TreeView.treeModel.SetData(MerinoCore.GetData());
 					m_TreeView.Reload();
 				}
 			}
@@ -372,7 +368,7 @@ namespace Merino
 					var newFile = AssetDatabase.LoadAssetAtPath<TextAsset>("Assets" + fullFilePath.Substring(Application.dataPath.Length));
 					AssetDatabase.ImportAsset(AssetDatabase.GetAssetPath(newFile));
 					LoadYarnFileAtFullPath(AssetDatabase.GetAssetPath(newFile), true);
-					m_TreeView.treeModel.SetData(GetData());
+					m_TreeView.treeModel.SetData(MerinoCore.GetData());
 					m_TreeView.Reload();
 
 					var fileID = MerinoData.FileToNodeID[newFile];
@@ -445,122 +441,6 @@ namespace Merino
 			}
 
 			return files;
-		}
-		
-		IList<MerinoTreeElement> GetDataFromFile(TextAsset source, int startID = 1)
-		{
-			var treeElements = new List<MerinoTreeElement>();
-			AssetDatabase.ImportAsset(AssetDatabase.GetAssetPath(source)); // TODO: only reload assets that need it? how to do that
-			//var format = YarnSpinnerLoader.GetFormatFromFileName(AssetDatabase.GetAssetPath(currentFile)); // TODO: add JSON and ByteCode support?
-			
-			// ROOT: create a root node for the file itself
-			var fileRoot = new MerinoTreeElement(source.name, 0, startID);
-			fileRoot.leafType = MerinoTreeElement.LeafType.File;
-			fileRoot.children = new List<TreeElement>();
-			treeElements.Add(fileRoot);
-			if (MerinoData.FileToNodeID.ContainsKey(source))
-			{
-				MerinoData.FileToNodeID[source] = startID;
-			}
-			else
-			{
-				MerinoData.FileToNodeID.Add(source, startID);
-			}
-
-			// load nodes
-			var nodes = YarnSpinnerLoader.GetNodesFromText(source.text, NodeFormat.Text);
-			var parents = new Dictionary<MerinoTreeElement, string>();
-			foreach (var node in nodes)
-			{
-				// clean some of the stuff to help prevent file corruption
-				string cleanName = MerinoUtils.CleanYarnField(node.title, true);
-				string cleanBody = MerinoUtils.CleanYarnField(node.body);
-				string cleanTags = MerinoUtils.CleanYarnField(node.tags, true);
-				string cleanParent = string.IsNullOrEmpty(node.parent) ? "" : MerinoUtils.CleanYarnField(node.parent, true);
-				
-				// write data to the objects
-				var newItem = new MerinoTreeElement( cleanName, 0, startID + treeElements.Count);
-				newItem.nodeBody = cleanBody;
-				newItem.nodePosition = new Vector2Int(node.position.x, node.position.y);
-				newItem.nodeTags = cleanTags;
-				if (string.IsNullOrEmpty(cleanParent) || cleanParent == "Root")
-				{
-					newItem.parent = fileRoot;
-					fileRoot.children.Add(newItem);
-				}
-				else
-				{
-					parents.Add(newItem, cleanParent); // we have to assign parents in a second pass later on, not right now
-				}
-				treeElements.Add(newItem);
-			}
-			
-			// second pass: now that all nodes have been created, we can finally assign parents
-			foreach (var kvp in parents )
-			{
-				var parent = treeElements.Where(x => x.name == kvp.Value).ToArray();
-				if (parent.Length == 0)
-				{
-					MerinoDebug.LogFormat(LoggingLevel.Error, "Merino couldn't assign parent for node {0}: can't find a parent called {1}", kvp.Key.name, kvp.Value);
-				}
-				else
-				{
-					// tell child about it's parent
-					kvp.Key.parent = parent[0];
-					// tell parent about it's child
-					if (kvp.Key.parent.children == null) // init parent's list of children if not already initialized
-					{
-						kvp.Key.parent.children = new List<TreeElement>();
-					}
-					kvp.Key.parent.children.Add(kvp.Key);
-				}
-			}
-			return treeElements;
-		}
-		
-		// FYI: TextAsset source basically does nothing right now, will be removed
-		IList<MerinoTreeElement> GetData (TextAsset source=null, bool isFromInit=false)
-		{
-			// init variables, create global tree root
-			var treeElements = new List<MerinoTreeElement>();
-			var root = new MerinoTreeElement("Root", -1, 0);
-			root.children = new List<TreeElement>();
-			treeElements.Add(root);
-			
-			// hack hack hack: if we're calling GetData from init, then ignore playmode currentFile
-			if (isFromInit && EditorApplication.isPlayingOrWillChangePlaymode)
-			{
-				source = null;
-			}
-
-			// ok, now let's load the data
-	
-			// then go through each file and get nodes for it, adding folder nodes as appropriate
-			int nodeID = 1;
-
-			foreach (var yarnFile in MerinoData.CurrentFiles)
-			{
-				// all folders are now created, let's add the yarn data now
-				var yarnData = GetDataFromFile(yarnFile, nodeID);
-				
-				// set the file node's parent to root
-				yarnData[0].parent = root;
-				root.children.Add(yarnData[0]);
-				
-				// add all data to tree elements
-				treeElements.AddRange( yarnData );
-				nodeID += yarnData.Count;
-			}
-			
-			// IMPORTANT: sort the treeElements by id!!!
-			treeElements = treeElements.OrderBy(x => x.id).ToList();
-			
-			// if there's already parent data then I don't really know what the depth value is used for (a cache to speed up GUI drawing?)
-			// but I think we're supposed to do this thing so let's do it
-			TreeElementUtility.UpdateDepthValues( root );
-			
-			MerinoData.TreeElements = treeElements;
-			return MerinoData.TreeElements;
 		}
 
 		#endregion
